Added an Interface for what a data repo should do, then moved the actual class to the implementation project.
Also renamed the projects and file names to be less confusing, and cleaned up the static reference files.  Probably should have done this as a separate commit so the Repo changes were a little more legible, but other than adding a couple methods to replace directly referencing the GameRepo.gameIndex, there were no changes to speak of.